+++ dotnet/Services/AffirmConstants.cs
@@ -26,7 +26,6 @@ namespace Affirm.Services
         public const string Void = "void";
         public const string Refund = "refund";
         public const string Update = "update";
-
         public const string AffirmUrlStub = "affirm.com";
         public const string AffirmApiVersion = "api/v1";
         public const string Charges = "charges";
@@ -43,6 +42,9 @@ namespace Affirm.Services
         public const string AlreadyCaptured = "already_captured";
 
         public const int MinimumDelayToCancel = 3600;
+        public const string HEADER_ACCEPT = "Accept";
+        public const string HEADER_CONTENT_TYPE = "Content-Type";
+        public const string PartialVoid = "partial_void";
 
         public class Inbound
         {
@@ -55,6 +57,16 @@ namespace Affirm.Services
             public const string Approved = "approved";
             public const string Denied = "denied";
             public const string Undefined = "undefined";
+            public const string VtexPaymentBaseUrl = "vtexpayments.com.br/api/pvt";
+            public const string Cancellations = "cancellations";
+            public const string AdditionalData = "additional-data";
+            public const string HEADER_VTEX_ACCOUNT = "X-Vtex-Account";
+            public const string HEADER_VTEX_API_APP_KEY = "X-VTEX-API-AppKey";
+            public const string HEADER_VTEX_API_APP_TOKEN = "X-VTEX-API-AppToken";
+            public const string VoidResponse = "voidResponse";
+            public const string ProxyAuthorization = "Proxy-Authorization";
+            public const string VtexUseHttps = "X-Vtex-Use-Https";
+            public const string VtexIdclientAutCookie = "VtexIdclientAutCookie";
         }
     }
 }